tena

+++ MyWebAPI/MyWebModels/ViewModels/BillVM.cs
@@ -11,11 +11,11 @@ namespace MyWebModels.ViewModels
         public int clientId { get; set; }
         public string clientName { get; set; }
         public double originalAmount { get; set; }
-        public int percentage { get; set; }
+        public double percentage { get; set; }
         public double totalAmount { get; set; }
         public double amountPaid { get; set; }
         public double restAmount { get; set; }
-        public int installmentCount { get; set; } // عدد الاقساط
+        public short installmentCount { get; set; } // عدد الاقساط
         public double delayFine { get; set; } = 0; // غرامه تاخير
         public InstallmentTypes delayFineType { get; set; }
         public InstallmentTypes installmentType { get; set; }
